add appbar stories
